[1008] - Finalização do exercício 1008.
    - Correção dos tipos das variáveis nos testes
    - Criação das regras de construção dos testes e documentação

+++ Beginner/1008/tests/src/salary.cpp
@@ -16,8 +16,8 @@ TEST_GROUP(salary)
     {
         int number;
         int hours;
-        float value;
-        float salary;
+        double value;
+        double salary;
         out *output;
     } test;
